make utils a pakage, delete brainhack data
