+++ small_TestData/README.md
@@ -1,0 +1,1 @@
+photos credit to: https://github.com/warmspringwinds/tensorflow_notes
